requirements.txt
# ── LLM & embeddings ──────────────────────────────────────────────────────────
openai>=1.0.0                  # OpenAI API client (GPT-4, embeddings)
anthropic>=0.25.0              # Claude API client (optional alternative LLM)
sentence-transformers>=2.6.0   # Local embedding models via Hugging Face

# ── Vector stores ─────────────────────────────────────────────────────────────
faiss-cpu>=1.7.4               # Local ANN index (swap for faiss-gpu if needed)
chromadb==1.5.2                # Embedded vector DB with persistence + HTTP client

# ── Audio transcription ───────────────────────────────────────────────────────
faster-whisper>=1.0.0          # CTranslate2 Whisper: ~4x faster than HF pipeline on CPU
mutagen>=1.47.0                # Audio duration metadata (optional but recommended)

# ── Document parsing ──────────────────────────────────────────────────────────
pypdf>=3.0.0                   # PDF text extraction
beautifulsoup4>=4.12.0         # HTML parsing
markdown>=3.5.0                # Markdown parsing

# ── API server ────────────────────────────────────────────────────────────────
fastapi>=0.110.0               # Web framework for the OpenAI-compatible API
uvicorn>=0.29.0                # ASGI server to run FastAPI

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0           # Load secrets from .env
tiktoken>=0.6.0                # Token counting for OpenAI models
tqdm>=4.66.0                   # Progress bars for batch operations
pydantic>=2.0.0                # Data validation / typed models

# ── Dev / testing ─────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
